feat(cli): pass CloudFormation parameters to "cdk deploy" (#6385)
* feat(aws-cdk): pass cloudformation parameters to "cdk deploy"

This commit closes #1237

I have added the command line option --parameters, -P which allows for
adding CloudFormation parameters at deploy time.

The functionality for deploying with parameters was already present,
but I have just added a way for the command line to pass it to the
deployment code.

The motivation for adding this functionality has been that we have
a setup where we share an account with access to CloudFormation, which
is why we do not want to have some fields in plain text, while they do
not have access to other services, which did not support secrets within
the fields we want to keep secret. This allows us to have parameters
that can be kept secret from the templates, while also having parameters
with NoEcho.

I have verified this functionality by both deploying a cdk project
with a CfnParameter from @aws-cdk/core, and an original CloudFormation
template with parameters.

* feat(cli): pass cloudformation parameter to "cdk deploy"

This changes the syntax of the --parameter parameter, so that it is now
possible to specify which stack a parameter is meant for, i.e. of you
want a parameter to apply to `MyStack`, then you would pass it as
--parameters MyStack:key=value
While still being able to pass parameters to all stacks by omitting the
stack name, i.e. --paramters key=value

This is being added to support wild card deployments, and deployments
with multiple stacks that do not neccessary contain the same parameters.

I have verified these changes by deploying a multistack cdk project,
where the stacks contains different parameters, and some shared
parameters.

* feat(cli): pass cloudformation parameters to "cdk deploy"

I have added some integration tests of both wildcard and single
deployments with parameters. While also fixing a bug found while
testing.

* feat(cli): pass cloudformation parameters to "cdk deploy"

Adding integration test for nested stacks.

Adding section in README.

* feat(cli): pass cloudformation parameters to "cdk deploy"

Adding an example of using parameters to the README.md

* update README

* more README cleanup

Co-authored-by: Elad Ben-Israel <benisrae@amazon.com>
Co-authored-by: Shiv Lakshminarayan <shivlaks@amazon.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>