feat(core): constructs library
This commit extracts the constructs programming model into an independent library called `constructs`. This library includes the Construct base class (and satellites) as well as the token system. The main motivation is to enable projects like [CDK for Kubernetes] to reuse these building blocks and allow interoperability between the frameworks (hence they have to use the same base types).

To ensure backwards compatibility until we release v2.0 of the AWS CDK, the `constructs-compat.ts` file implements shadow and wrapper types for types that are now in "constructs" so the AWS CDK code remains unchanged and fully passes API compatibility tests.

The token system and a bunch of types related to dependencies has been _duplicated_ since the effort of shadowing was too great and there is no major value (it would still be possible to interoperate even if we have two token systems).

[CDK for Kubernetes]: https://github.com/awslabs/cdk8s