chore(eks): consolidate integ tests (#6636)
chore(eks): consolidate integ tests (#6636)

Since creating a VPC and an EKS cluster takes about ~30 minutes and integration tests had to be executed for any change in one of the resource providers' code, the EKS integration tests became a major barrier for contributions and evolution of the EKS module.

Pragmatically, this change consolidates all "kubectl-enabled" tests into a single integration test which exercises all the relevant features. This is not perfect, but at least will allow us to evolve.