chore: add missing ECR permissions to publishing role
The bootstrapped publishing role had permissions to write to ECR,
but not actually to check what was already uploaded (to be able to
skip double work).

`BatchCheckLayerAvailability` is necessary do to pushes at all.

Plus, it needs permissions to get a token to pass to `docker login`,
which *must* be on `resource: *`.