feat(apigateway): rate limited API key (#6509)
In cases when we need to configure throttling and quotas etc., at the api key level, we need to create a usage plan for the api key and link the two.
This commit introduces a construct which allows the user to create an api key and spicify rate limiting settings, while creating the api key and not worry about creating usage plan and linking the two.

fixes #6405