refactor(ec2): MachineImages create appropriate UserData
Previously, MachineImages were dumb objects that maybe had UserData
given by the user, and maybe not. If the UserData object was absent,
downstream objects had to inspect the OS type of the MachineImage to
create the right type of UserData object.

That is pretty silly--polymorphism should take care of that unnecessary
`if` there. The MachineImage knows what type of OS it represents, and
should be the one generating the right type of UserData.

Move the responsibility from `Instance` and `AutoScalingGroup` to
`MachineImage`.

BREAKING CHANGE: if you implemented a custom subclass of `IMachineImage` it must now always return a `userData` object.