feat(core): the "constructs" module (#6623)
feat(core): the "constructs" module (#6623)

This commit extracts the constructs programming model into an independent library called `constructs` ([npm], [Maven], [PyPI], [NuGet]). This library includes the `Construct` base class (and satellites) as well as the token system. The main motivation is to enable projects like [CDK for Kubernetes] to reuse these building blocks and allow interoperability between the frameworks (hence they have to use the same base types).

To ensure backwards compatibility until we release v2.0 of the AWS CDK, the `construct-compat.ts` file implements shadow and wrapper types for types that are now in "constructs" so the AWS CDK code remains unchanged and fully passes API compatibility tests.

The token system and a bunch of types related to dependencies has been _duplicated_ since the effort of shadowing was too great and there is no major value (it would still be possible to interoperate even if we have two token systems).

Added an additional dependency on `constructs` to all AWS CDK modules.

MONOCDK: has been modified to take a `peerDependency` on this new module as this will be the final setup (users will have to depend on both `monocdk` and `constructs`), so we are not going to take shortcuts here.

TESTING: All relevant unit tests were *duplicated* from **@aws-cdk/core** to **constructs**, and intentionally maintained unchanged in core to ensure backwards compatibility.

[npm]: https://www.npmjs.com/package/constructs
[Maven]: https://search.maven.org/search?q=g:software.constructs
[PyPI]: https://pypi.org/project/constructs
[NuGet]: https://www.nuget.org/packages/Constructs/
[CDK for Kubernetes]: https://github.com/awslabs/cdk8s