rfc(bootstrap): bring back the removed `cdk bootstrap` CLI options (#5757)
In the original RFC for the modified cdk bootstrap command done in #4461,
we removed two options from the cdk bootstrap command-line utility:
`--toolkit-bucket-name` and `--bootstrap-kms-key-id`.
In this small revision to that RFC,
I propose we not remove them to preserve backwards compatibility for customers who already use those options,
and I show what changes that would require to our bootstrap CloudFormation template.