fix(cloudwatch): unhelpful error when reusing metric IDs (#6892)
Reuse of the same metric ID was already checked upon construction of the
`MathExpression` instance and so we thought this error couldn't occur.
However, reuse can also occur (and is also prohibited) between different
`MathExpression`s, even on different axes.

Make the error message describe the situation in a way that is
actionable.