fix(ecr-assets): Docker asset path regression
Previously, users were allowed to pass a the `file` argument in
DockerImageAssetProps with a directory included in the string, IE:
'my-dir/Dockerfile'. This worked because when calling `docker build`,
the --file argument was formatted by prepending the asset's staging
directory. However, this was removed in v1.30.0 with the assumption that
`file` could contain only a filename and `directory` included all
other relative path information. This resulted in the docker build
command receiving a --file argument that was relative to the ${cwd} and
possibly wrong.

This is resolved by using `path.resolve` to get the file's full path.
Changes to how we are passing directory information to compute the
asset's hash were required to not break existing asset hashes. Otherwise
there is a possibility that a directory other than the one where the
Dockerfile resides was being referenced when hashing the asset's
contents.

fixes: #6814