Handle conflicting operators and keys
In the case of both conflicting keys and conflicting operators,
the behaviour is inherited from the way that
`PolicyStatement.addCondition` already works.

When the same operator is specified in two different conditions,
the objects will be merged.

When the same key is specified under the same operator in two
different conditions, the one added later will overwrite the
existing one.