chore: mergify to only dismiss approvals on stale PRs (#6974)
Currently, mergify dimisses all reviews on a PR when new commits are
added. Users, quite often, continuously push commits to a PR as they get
through all of the requested changes.
This means that the PR moves from 'changes reqeusted' state to 'review
required' state even though it's not ready.

Change this workflow so that only approvals are dismissed when PRs are
modified after an approval. Further, add a comment to let the user know
that they should dismiss the review when the PR is ready for review.