chore(cognito): integ tests for sign up (#6970)
* chore(cognito): integ tests for sign up

Configuring the correct properties for sign up and testing it is hard.

This PR - https://github.com/aws/aws-cdk/pull/6938 - shows how it can be
easily mis-configured.

Testing the correct configuration is hard and requires a certain amount
of set up and the correct set of calls to the 'SignUp' API.

The integ tests here encode all of this in a way that can be easily
replicated when any part of sign up is being modified or needs to be
tested.

* fixup breaking tests

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>