feat(codebuild): expose aws/windows/base:2.0 image (#7004)
The 1.0 version of this image, while still supported by the service, is
no longer documented on the CodeBuild supported images documentation.
Users are recommended to upgrade to the 2.0 version of the image, which
also comes with up-to-date versions of Node, DotNet SDK, etc...

This PR makes a new `WindowsBuildImage.WINDOWS_BASE_2_0` constant
available to easily access this build image, and proceeds to deprecate
the older `WindowsBuildImage.WIN_SERVER_CORE_2016_BASE` image.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>