fix(cli): can't use credential providers for stacks with assets (#7022)
* fix(cli): can't use credential providers for stacks with assets

Stacks with assets were broken for environments that completely
used credential providers (and no other sources of AWS credentials).

Not because the wrong credentials were being used (as you would expect
it to be broken). Instead, an error was being thrown because the
"current" account lookup was incorrectly forwarding to the "default"
account lookup (fix: we know what the current account is. Just return
that instead).

What's worse, this value wasn't even being used! It was being looked up
so that if something was wrong finding the target bucket, we could
format a nice error message containing the account id. Even happy
paths were failing due to the premature lookup though (fix: only
look up when we actually need the value).

Fixes #7005.

* Properly look up partition so this also works in aws-cn

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>