chore: revert mergify behaviour (#7025)
A [previous commit][1] changed the behaviour of mergify to
not dismiss PR reviews marked 'changes requested', and let
the PR authors dismiss the review when they're ready.

It was not considered that dismiss PR reviews is an privileged
action available only to admins of the repo. Quoting [Github
documentation][2], "repository administrators or people with
write access can dismiss a review".

Reverting this change to its previous behaviour.

[1]: https://github.com/aws/aws-cdk/commit/9ad04485021a3132dd2bb97475fb4f945df7df3a
[2]: https://help.github.com/en/github/collaborating-with-issues-and-pull-requests/dismissing-a-pull-request-review