chore(cli): fix bootstrap v2 by adding a missing 'await' (#7087)
The title says it all.

Q: Why wasn't this caught in testing/why didn't you add a test?

A: We had tests on this originally, but we reverted them because
they made assumptions on the source environment and would fail
when run in the pipeline. Restoring those is next on the agenda.

https://github.com/aws/aws-cdk/commit/2e9f241833ea20112c134eaa904425a26007e9d6