feat(cognito): import an existing user pool client
Include other clean up around awslint exclusions and CloudFormation
resource attributes.

BREAKING CHANGE: `UserPoolClient` construct no longer has the property
`userPoolClientClientSecret`. The functionality to retrieve the client
secret never existed in CloudFormation, so this property was not
working in the first place.
* **cognito:** The `userPoolClientName` property on the `UserPoolClient`
construct may now return `undefined`. This property was previously
incorrectly configured and was returning a not-implemented message from
CloudFormation. It now returns the client name if one was configured by
the user, `undefined` otherwise.