fix(core): the Token API conflicts with constructs
Both the `@aws-cdk/core` and `constructs` libraries defined their
version of the `Token` API, and the `TokenMap` was stored in a global
variable that was the same for both. As a result, the first loaded
library (`constructs`) would "win". While this isn't a problem in the
TypeScript world, it breaks statically typed languages such as Java and
C# that do not support structural typing.

This changes how the `@aws-cdk/core` API is implemented so that it
wraps the `constructs` features instead of duplicating them. This
involves wrapping types to preserve backwards compatibility of the
old(er) code built against the `@aws-cdk/core` API (particularly that
written in Java or C#).

Fixes #6978