feat(kinesis): streams are encrypted with the KMS master key by default
Currently, the default for a stream is to be unencrypted. This change uses
the KMS master key and enables encryption on Streams by default.

Stream encryption is not supported in `cn-north-1` and `cn-northwest-1` regions
so there's a condition that sets the default in those regions to unencrypted.

Why is this not a breaking change:
Modifying the encryption settings on a Stream does not require interruption.
Adding/removing the `StreamEncryption` properties will enable/disable the encryption
on the stream accordingly without interruption.