Address PR comments
- Change the conditions type to be `Record<string, Record<string, any>>`
so as not to cause breaking changes once JSII transpiles. The change
from an interface to a type alias is necessary because an interface in TS
will convert to a named type in languages with nominal rather than
structural typing. The change from `string | string[]` to `any` is needed
because union types can't be supported.
- Rather than copy the conditions from the wrapped principal when the
`PrincipalWithConditions` is constructed, just merge the conditions in the
getter for `conditions` to avoid any problems with tokens.
- Add documentation for all public members.
- Make wrapped principal private.
- Inherit `assumeRoleAction` from the wrapped principal rather than always
using "sts:AssumeRole".