chore: security diff calculation fails for nonexistent stack
Before the refactor, looking up the template for a nonexistent
stack returned `{}`, an empty template.

During the refactor, this was changed to throw an error, which
is why deployments of new stacks started failing (the lookup
of the template for the purposes of the security diff would fail).

Restore the original behavior.