fix(ecs): default Service throws in a VPC without private subnets (#7188)
People like to create VPCs without private subnets, as the NAT
gateways (or NAT instances) can be expensive.

The default settings of an ECS Service is to create themselves
in a PRIVATE subnet, instead of the "default selection order" of
Private -> Isolated -> Public. Now use the default selection
(by not supplying a `subnetType` at all).

Fixes #7062.