feat(bootstrap): require `aws:SecureTransport` for staging bucket
Many of our customers are going to run the open source tool `Scout2`
over their AWS accounts to determine their security posture.

By default, this tool is going to report the lack of an
`aws:SecureTransport` restriction on S3 buckets with "medium" severity,
which will alarm users and for those in enterprise environments
require them to address the situation.

The bootstrap stack is one which users don't have a lot of control over
(without going to a full-fledged custom template), and enabling this
flag out of the box seems like a good default, so that the most common
posture-checking tool won't trip over a default CDK setup.