fix(lambda): circular reference when Cognito permissions are added to function role
Lambda functions can be used as [triggers] on Cognito user pools. It is
common for these triggers to query (callback) the user pool for more
information.

The function's IAM policy needs to updated with permissions to call the
Cognito IDP service with the resource restricted to the specific user
pool. Adding such an IAM policy creates a circular reference between the
user pool, which refers to the lambda function, and the function policy,
which refers to the user pool.

This fixes the problem by using the `functionName` initialization
property, when specified, to determine the `functionArn`, thereby,
removing its dependency on the function being created first.

fixes #7016

[triggers]: https://docs.aws.amazon.com/cognito/latest/developerguide/cognito-user-identity-pools-working-with-aws-lambda-triggers.html