fix(autoscaling): adds noDevice as a volume type
fixes #7242

Currently it's impossible to disable a mapping because the volume property is
required on a BlockDevice object. CloudFormation will fail at deploy time
with a validation failure because NoDevice can not be present at the same time
as Ebs or VirtualName. This adds a noDevice volume that can be set
on the BlockDevice to supress a drive mapping.

wip