fix: new IAM Condition type is unusable in Java (#7270)
The changing of the type of `Condition` from `any` to `Record<string,
any>` broke common code in Java.

Many users would be passing a `Map<String, String>` in that location,
which WOULD be assignable to the old type `Object`, but not to the
new type `Map<String, Object>`.

Revert for now and turn this into a feature request to jsii team.