fix(cli): --app command does not work when executing a command without arguments (#7249)
If the application passed with the `--app` parameter is executable, we were failing on a `fs.stat` call as it was not a file that was provided. This change swallows up that error and returns the provided command without modification.

Also added some tests (thank you rix0rrr@ for getting this started) and some other minor improvements.

Closes #6930