fix(ecs): using secret JSON field with fargate task does not fail
For tasks that use the Fargate launch type, it is only supported to inject
the full contents of a secret as an environment variable. Specifying a specific
JSON key or version is not supported at this time.

Also clean up/refactor a bit.

See https://docs.aws.amazon.com/AmazonECS/latest/userguide/specifying-sensitive-data-secrets.html#secrets-considerations

Closes #7272