fix(eks): unable to create KubernetesResources in another stack (#7322)
kubectl resources such as KubernetesResource and HelmCharts require a custom resource provider. We used to create the resource provider in the stack in which the k8s resource was defined but this resulted in a cyclic reference since the eks.Cluster needs to reference the resource provider and the k8s resource needs to reference the cluster.

The solution is that the kubectl resource provider is always created in the stack in which the cluster is defined. This makes sure that references always go in one direction (and also makes much more sense).

Fixes #7231