feat(ec2): filtering selected subnets by availability zone
This allows a user to specify that their subnet selection should only return subnets that are in the specified availability zones. I added this to provide a native solution to the problem of VPC endpoint services not existing in all AZs. The way CDK currently creates VPC endpoints, it places one in every AZ, which for some services results in this CloudFormation error on deployment:

> The VPC endpoint service <-service name-> does not support the availability zone of the subnet

CDK cannot currently know the subnets in advance, so I added functionality to SubnetSelection to allow for choosing the availability zones, like so:

```ts
new InterfaceVpcEndpoint(stack, "VPC Endpoint", {
  vpc,
  service: new InterfaceVpcEndpointService("com.amazonaws.vpce.us-east-1.vpce-svc-uuddlrlrbastrtsvc", 443),
  subnets: {
    subnetType: SubnetType.ISOLATED,
    availabilityZones: ["us-east-1a", "us-east-1c"]
  }
```

This provides the user a way to solve this problem, because they can get the endpoint service availability zones from the AWS console.