feat(kinesis): add `grant` API to IStream
The `grant` API enables the capability to provide permissions on a stream
to an IAM principal.

Context:
`DescribeStream` permissions are redundant with `DescribeStreamSummary`
and read operations do not need permissions to both. However, there are
cases (i.e. Lambda::EventSourceMapping ) where the presence of the
permission is being validated when creating the resource.

This change exposes the `grant` API on an IStream so that the necessary
permissions can be provided with an extra API call

BREAKING CHANGE:
`grantRead()` API  no longer provides permissions to `kinesis:DescribeStream`.
Use `kinesis:DescribeStreamSummary` in it's place or add it if desired through the `grant()` API