chore(cli): fix wildcard deploy with parameters integ test
we had an integ test that does a deploy of all stacks prefixed `${STACK_NAME_PREFIX}-param-test-\*`. Another stack that matches that pattern was added recently and this integ test was not updated to provide parameters. 

When running the test CloudFormation complains that parameters are not provided when they're expected. This change adds in the expected parameters for the new stack.