feat: update "constructs" to 3.x (#7408)
`constructs@2.x` included the token system which could cause conflicts with the token system we already have in @aws-cdk/core, so we decided to remove it for now from `constructs` and introduce at a later stage.

This change uncovered a couple of new, unreleased, changes (kinesis and core/refs) which accidentally took a dependency on `IResolvable` from `constructs` instead of `@aws-cdk/core`, which might have created issues in strong-typed languages like Java and .NET, so this PR must be merged before the next release.

Fixes #6978