fix(glue): DataFormat constants are not visible in non-JS languages (#7458)
The constants on `DataFormat` were modelled as static properties on an
interface, which is an invalid pattern in `jsii`. The `export const`
declarations were in fact ignored by `jsii` at compilation, and the
`namespace` declaration causes a conflict with the interface itself when
the submodule feature is enabled on the codebase.