fix(cli): can't bootstrap environment not in app
It used to be that if we had an `--app` argument, we would always glob
arguments to `cdk bootstrap <ENV>` through the environments of stacks
in the app.

This makes it super hard/annoying to run `cdk bootstrap
aws://1234/us-somewhere` in a CI/CD project; you have to add a stack
there first and compile before you're allowed to do that, which is
kinda silly.

Change behavior to only glob environment from the environments in the
app if it looks like the user is supplying a glob (if it contains `*`).

If the user is supplying a concrete environment name, just use it
directly.

Also in this commit:

- refactor: lots of places where we were passing around a pair
  of `(account, region)`. Replace those by passing a `cxapi.Environment`
  instead (most of the changes in this PR).
- refactor: the old and new bootstrapping experience had a lot of
  copy/pasta between them. Refactored to make them share code.
- (prerelease) feat: add a version check to the bootstrapping operation,
  so that users won't accidentally downgrade a bootstrap stack to an older
  version (This happened to 2 devs already, and is an easy mistake to
  make. Protect against it.)