fix(eks): version update completes prematurely
The `UpdateClusterVersion` operation takes a while to begin and until then, the cluster's status is still `ACTIVE` instead `UPDATING` as expected. This causes the `isComplete` handler, which is called immediately, to think that the operation is complete, when it hasn't even began.

Add logic to the cluster version update `onEvent` method to wait up to 5 minutes until the cluster status is no longer `ACTIVE`, so that the subsequent `isComplete` query will be based on the version update operation itself.

Extended the timeout of `onEvent` to 15m to ensure it does not interrupt the operation.

TESTING: Updated unit tests to verify this retry behavior and performed a manual upgrade tests while examining the logs.

Fixes #7457