fix(assets): infrequent "ValidationError: S3 error: Access Denied" (#7556)
We're checking if an asset exists in S3 before uploading (to avoid
re-uploading possibly large assets). This causes HEAD/GET-after-PUT to
become eventually consistent because it creates a negative cache entry!

This switches to using a LIST-based check, which will not create a
negative cache entry, and preserve the immediate consistency of
GET-after-PUT in this workflow.

BREAKING CHANGE: `cdk deploy` now needs `s3:ListBucket` instead of `s3:HeadObject`.

Fixes #6430
Fixes #7553