fix(cli): context keys specified in `cdk.json` get moved to `cdk.context.json
Removes migration logic that was introduced to move legacy context provider keys.

Rationale:
This logic was intended to be removed prior to v1.0
It looks for any key that contains a `:` and moves it to `cdk.context.json`. This is
not expected behavior and also prevents users from having keys that have `:`. Our
init templates also include feature flags which make use of the `:` character.

Closes #7399