fix(stepfunctions-tasks): cannot specify part of execution data or task context as input to the `RunLambda` service integration (#7428)
The Lambda service integration requires an input field called `payload`.
We modeled this as a `{[key: string]: any}` which precludes the usage of
execution data or context data as inputs to a Lambda function.

Fix:
Change the type of `payload` to be `TaskInput` which has already modeled
a type union for task classes that accept multiple types of payload.
This will enable usage of literal strings, objects, execution data, and
task context as input types supported for invoking a Lambda function.

Rationale:
Although this is a breaking change, the workarounds for enabling usage of
different types is not user friendly and incomplete as all of the types
above cannot be expressed in the current modeling of `payload`

Fixes #7371

BREAKING CHANGE:
`payload` in RunLambdaTask is now of type `TaskInput` and has a default of the state input instead of the empty object.
You can migrate your current assignment to payload by supplying it to the `TaskInput.fromObject()` API