docs(cli): expand help for --bootstrap-bucket-name option (#7645)
Changes the help for --bootstrap-bucket-name flag (for bootstrap subcommand) to note that the bucket will be created and must not exist.

This also incidentally adds a space in the left column so it looks less like `--toolkit-bucket-name` is its own flag (with no help, only [string[) which was a recent customer issue.

Before:
```
--bootstrap-bucket-name, -b,              The name of the CDK toolkit bucket
--toolkit-bucket-name                                                 [string]
```

After:

```
--bootstrap-bucket-name, -b,              The name of the CDK toolkit bucket;
--toolkit-bucket-name                     bucket will be created and must not
exist                       [string]
```